# StreamSyncBackend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSync API
  version: 1.0.0
  description: Real-time collaborative workspace backend. Workspaces, projects, documents,
    tasks, activity and AI assistance.
paths:
...
  /api/auth/register/:
    post:
      operationId: auth_register
      description: Creates an account and signs the new user in. Passwords are checked
        against Django's configured validators; failures come back as field errors
        on `password`.
      summary: Create an account
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                Session:
                  value:
                    user:
                      id: 6f1b7c58-6d3a-4a5e-9e5e-9c0b1f2a3d4e
                      name: Raj Kumar
                      email: raj@streamsync.app
                      avatar_url: null
                      title: null
                      created_at: '2026-08-03T10:15:00Z'
                    access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    expires_at: '2026-08-03T10:30:00Z'
          description: Authenticated. The refresh token is set as an httpOnly cookie
            and is deliberately absent from this body.
        '400':
          description: Validation error, e.g. email already registered.
        '429':
          description: Too many registration attempts.
  /api/auth/login/:
    post:
      operationId: auth_login
      description: Exchanges email and password for a token pair. Wrong credentials
        and unknown accounts return an identical 401, so this endpoint cannot be used
        to discover which emails are registered.
      summary: Sign in
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                Session:
                  value:
                    user:
                      id: 6f1b7c58-6d3a-4a5e-9e5e-9c0b1f2a3d4e
                      name: Raj Kumar
                      email: raj@streamsync.app
                      avatar_url: null
                      title: null
                      created_at: '2026-08-03T10:15:00Z'
                    access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    expires_at: '2026-08-03T10:30:00Z'
          description: Authenticated. The refresh token is set as an httpOnly cookie
            and is deliberately absent from this body.
        '400':
          description: Malformed request body.
        '401':
          description: Incorrect email or password.
        '429':
          description: Too many sign-in attempts.
  /api/auth/google/:
    post:
      operationId: auth_google
      description: Exchanges a verified Google ID token — the `credential` Google's
        Sign In With Google button hands the frontend — for a StreamSync session.
        The first sign-in for a given Google identity creates the account automatically;
        an existing password account sharing the same verified email is linked rather
        than duplicated.
      summary: Sign in with Google
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                Session:
                  value:
                    user:
                      id: 6f1b7c58-6d3a-4a5e-9e5e-9c0b1f2a3d4e
                      name: Raj Kumar
                      email: raj@streamsync.app
                      avatar_url: null
                      title: null
                      created_at: '2026-08-03T10:15:00Z'
                    access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    expires_at: '2026-08-03T10:30:00Z'
          description: Authenticated. The refresh token is set as an httpOnly cookie
            and is deliberately absent from this body.
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                Session:
                  value:
                    user:
                      id: 6f1b7c58-6d3a-4a5e-9e5e-9c0b1f2a3d4e
                      name: Raj Kumar
                      email: raj@streamsync.app
                      avatar_url: null
                      title: null
                      created_at: '2026-08-03T10:15:00Z'
                    access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    expires_at: '2026-08-03T10:30:00Z'
          description: Authenticated. The refresh token is set as an httpOnly cookie
            and is deliberately absent from this body.
        '401':
          description: The Google credential could not be verified.
        '429':
          description: Too many attempts.
        '503':
          description: Google sign-in is not configured.
  /api/auth/refresh/:
    post:
      operationId: auth_refresh
      description: Reads the refresh token from its httpOnly cookie, revokes it and
        issues a new pair. The request body is ignored. Returns the user as well as
        the access token so a page reload can restore the session in one round trip.
      summary: Refresh the session
      tags:
      - auth
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                Session:
                  value:
                    user:
                      id: 6f1b7c58-6d3a-4a5e-9e5e-9c0b1f2a3d4e
                      name: Raj Kumar
                      email: raj@streamsync.app
                      avatar_url: null
                      title: null
                      created_at: '2026-08-03T10:15:00Z'
                    access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    expires_at: '2026-08-03T10:30:00Z'
          description: Authenticated. The refresh token is set as an httpOnly cookie
            and is deliberately absent from this body.
        '401':
          description: Missing, expired or already-used refresh token.
  /api/auth/logout/:
    post:
      operationId: auth_logout
      description: 'Blacklists the refresh token and clears its cookie. Idempotent:
        returns 204 even when there was no session to end.'
      summary: Sign out
      tags:
      - auth
      security:
      - {}
      responses:
        '204':
          description: Signed out.
  /api/auth/me/:
    get:
      operationId: auth_me
      description: |-
        The signed-in user.

        Uses the project-wide default permission (IsAuthenticated), so it is also
        the smallest possible check that JWT authentication is wired up correctly.
        Stated explicitly rather than inherited, because the security of this
        endpoint should be readable without opening the settings module.
      summary: Current user
      tags:
      - auth
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
        '401':
          description: No or invalid access token.
...
components:
  schemas:
...
    GoogleLoginRequest:
      type: object
      description: |-
        Google sign-in input.

        `credential` matches the field name Google's own Identity Services library
        hands the frontend's callback — keeping it means the client forwards the
        button's payload unchanged rather than renaming a key for no reason. It is
        the signed ID token JWT, verified server-side in `services.py`; nothing
        about it is trusted until that verification passes.
      properties:
        credential:
          type: string
          minLength: 1
        remember_me:
          type: boolean
          default: false
      required:
      - credential
...
    LoginRequest:
      type: object
      description: |-
        Login input.

        Credentials are only checked for shape here. Verifying them is
        authentication, which belongs to the service layer.
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 128
        remember_me:
          type: boolean
          default: false
      required:
      - email
      - password
...
    RegisterRequest:
      type: object
      description: |-
        Registration input.

        A plain Serializer rather than a ModelSerializer: a ModelSerializer would
        expose every writable model field by default, so `is_staff` would be one
        forgotten line away from being settable by an anonymous request.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 128
      required:
      - email
      - name
      - password
...
    Session:
      type: object
      description: |-
        What every successful authentication returns.

        The refresh token is deliberately absent: it travels only as an httpOnly
        cookie. Putting it in the body would hand it to any script on the page and
        defeat the reason for the cookie. (README §25)
      properties:
        user:
          allOf:
          - $ref: '#/components/schemas/User'
          readOnly: true
        access:
          type: string
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - access
      - expires_at
      - user
...
    User:
      type: object
      description: |-
        The public shape of a user.

        Explicit `fields` rather than `exclude`: a field added to the model must
        never appear in an API response merely because someone forgot to hide it.
        That is how password hashes and internal flags leak.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
          title: Full name
          description: Display name shown to collaborators.
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
          description: Used to sign in. Case-insensitively unique.
        avatar_url:
          type: string
          readOnly: true
        title:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - avatar_url
      - created_at
      - email
      - id
      - name
      - title
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT